issue: 418439 runEventLoop returned after wrong number of processed events (not maxEvents as requested by applciation) this cause runBreakEventLoop test to fail on wrong timeout values
TEARDOWN event, which is not provided to user, trigger exiting from runEventLoop too early.
In runBreakEventLoop test we loop between sleeping on timeout and sleeping with bad CleintSession.
The CS failure had an addition internal JXIO event (TEARDOWN) pending in the queue
which caused the loop to exit and measure wrong timeout values and fail the test.

Change-Id: I90d35dc34c52fea37051d19edc1fefa309683783